Switched to new Algolia account
Algolia is changing the way they do DocSearch, instead of it being
something that they manage and we have to email them to get
configuration changes, they now give us free Algolia accounts so we can
configure it ourselves. I've updated the search to use the new account.
Currently, only I have access to this account I think, but I guess I can
add other people.

I also fixed a bug - it was only searching the current docs, this change
means it now searches the version that the user is currently viewing.